+++ config/max-lines-baseline.txt
@@ -10,7 +10,6 @@ extensions/amazon-bedrock/stream.runtime.ts
 extensions/anthropic/index.test.ts
 extensions/anthropic/register.runtime.ts
 extensions/anthropic/session-catalog.test.ts
-extensions/anthropic/session-catalog.ts
 extensions/browser/src/browser-tool.test.ts
 extensions/browser/src/browser-tool.ts
 extensions/browser/src/browser/cdp.ts
